Refactor loop cleanup to the common util.TestCase class
